backend: Implement partial SDK ScriptUnpacking workaround
This workaround only works partially where WE have full control over the
Script fields that are unpacked. After using this fix we had to realize
that there is a hidden call further down the callstack in the SDK.

We will have to submit a PR for this to get fixed. In the meantime we
will update using a small patch on the perun fork for the ckb-sdk-go.